+++ src/components/BoardContent/index.stories.tsx
