+++ src/components/BoardContent/style.ts
